requirements.txt
discord.py>=0.16.8
aiohttp
websockets>=3.1
chardet
beautifulsoup4
